Move the sources in a seperate file
